fix bug 1073473 - Move search app into kuma namespace.
