+++ kuma/search/management/commands/reindex.py
@@ -3,7 +3,7 @@ from optparse import make_option
 
 from django.core.management.base import BaseCommand, CommandError
 
-from search.commands import es_reindex_cmd
+from ...commands import es_reindex_cmd
 
 
 class Command(BaseCommand):
