+++ kuma/search/signals.py
