+++ kuma/search/templates/search/plugin.html
