+++ kuma/search/tests/test_plugin.py
