+++ kuma/search/urls.py
@@ -2,11 +2,11 @@ from django.conf.urls.defaults import patterns, url
 from rest_framework.urlpatterns import format_suffix_patterns
 
 # this allows using ".json" extensions for the view to force json output
-urlpatterns = format_suffix_patterns(patterns('search.views',
+urlpatterns = format_suffix_patterns(patterns('kuma.search.views',
     url(r'^$', 'search', name='search'),
 ))
 
-urlpatterns += patterns('search.views',
+urlpatterns += patterns('kuma.search.views',
     url(r'^/xml$', 'plugin', name='search.plugin'),
     url(r'^/suggestions$', 'suggestions', name='search.suggestions'),
 )
