+++ urls.py
@@ -36,7 +36,7 @@ urlpatterns = patterns('',
     (r'^admin/', include('smuggler.urls')),
     (r'^admin/', include(admin.site.urls)),
 
-    (r'^search', include('search.urls')),
+    (r'^search', include('kuma.search.urls')),
 
     # Special-case here because this used to live in the wiki app and
     # needs to keep its historical URL.
